+++ changelog.md
@@ -3,6 +3,11 @@ This file contains all the notable changes done to the Ballerina Kafka package t
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Removed
+- [Removed deprecated `kafka:ConsumerRecord` and `kafka:ProducerRecord`](https://github.com/ballerina-platform/ballerina-library/issues/6359)
+
 ## [3.10.2] - 2024-03-28
 
 ### Fixed